Compiler warnings disable some compiler warnings of VS2015 compiler

+++ test/common/HashTest.cpp
@@ -42,7 +42,7 @@ TEST_F( HashTest, CreateTest ) {
     EXPECT_EQ( myHash2.hashValue(), 10 );
 
     Hash myHash3( "test", 7 );
-    EXPECT_NE( myHash3.hashValue(), 0 );
+    EXPECT_NE( myHash3.hashValue(), (unsigned int) 0 );
 }
 
 //---------------------------------------------------------------------------------------------
@@ -53,18 +53,18 @@ TEST_F( HashTest, MakeStringHashTest ) {
     
     std::string value;
     value = ( "huhu1" );
-    const int hash1 = Hash::toHash( value.c_str(), Base );
-    EXPECT_NE( hash1, 0 );
+    const unsigned int hash1 = Hash::toHash( value.c_str(), Base );
+    EXPECT_NE( hash1, ( unsigned int ) 0 );
     EXPECT_LE( hash1, Base );
 
     value = ( "huhu2" );
-    const int hash2 = Hash::toHash( value.c_str(), Base );
-    EXPECT_NE( hash2, 0 );
+    const unsigned int hash2 = Hash::toHash( value.c_str(), Base );
+    EXPECT_NE( hash2, ( unsigned int ) 0 );
     EXPECT_LE( hash2, Base );
 
     value = ( "huhu3" );
-    const int hash3 = Hash::toHash( value.c_str(), Base );
-    EXPECT_NE( hash3, 0 );
+    const unsigned int hash3 = Hash::toHash( value.c_str(), Base );
+    EXPECT_NE( hash3, ( unsigned int ) 0 );
     EXPECT_LE( hash3, Base );
 
     Hash myHash_inited( value.c_str(), Base );
@@ -77,18 +77,18 @@ TEST_F( HashTest, MakeUIntHashTest ) {
     EXPECT_EQ( myHash_empty.hashValue(), 0 );
 
     unsigned int value = 17;
-    const int hash1 = Hash::toHash( value, Base );
-    EXPECT_NE( hash1, 0 );
+    const unsigned int hash1 = Hash::toHash( value, Base );
+    EXPECT_NE( hash1, ( unsigned int ) 0 );
     EXPECT_LE( hash1, Base );
 
     value = 27;
-    const int hash2 = Hash::toHash( value, Base );
-    EXPECT_NE( hash2, 0 );
+    const unsigned int hash2 = Hash::toHash( value, Base );
+    EXPECT_NE( hash2, ( unsigned int ) 0 );
     EXPECT_LE( hash2, Base );
 
     value = 37;
-    const int hash3 = Hash::toHash( value, Base );
-    EXPECT_NE( hash3, 0 );
+    const unsigned int hash3 = Hash::toHash( value, Base );
+    EXPECT_NE( hash3, ( unsigned int ) 0 );
     EXPECT_LE( hash3, Base );
 
     Hash myHash_inited( value, Base );
